TIP-842: Move association type normalizer to structure sub context
