+++ src/Akeneo/Pim/Structure/Component/Normalizer/InternalApi/AssociationTypeNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Normalizer;
+namespace Akeneo\Pim\Structure\Component\Normalizer\InternalApi;
 
 use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Pim\Bundle\VersioningBundle\Manager\VersionManager;
