+++ src/Akeneo/Pim/Structure/Component/Normalizer/Standard/AssociationTypeNormalizer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Standard;
+namespace Akeneo\Pim\Structure\Component\Normalizer\Standard;
 
 use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
+use Pim\Component\Catalog\Normalizer\Standard\TranslationNormalizer;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
