+++ src/Akeneo/Pim/Structure/Component/spec/Normalizer/InternalApi/AssociationTypeNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Normalizer;
+namespace spec\Akeneo\Pim\Structure\Component\Normalizer\InternalApi;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\VersioningBundle\Manager\VersionManager;
